Initial test implementation to draw map annotation text always on the minimap.
Much better looking text - updated code flow so that text is always on top of icons.

Added a dedicated font (with ability in the UI to set it) for map annotations. Map annotation's view rect is now padded slightly to account for the drop shadow. Map annotations will now stop displaying when sufficiently zoomed out.